refactor: move definition of RegexLiteral to make use in evaluator
